categori id FK olarak oluşturuldu, create-update-delete operasyonlarına gerekli koşullar eklendi

+++ CrudOperations.Business/Concrete/CategoryService.cs
@@ -4,7 +4,10 @@ using CrudOperations.DataAccess.Concrete;
 using CrudOperations.Entities;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Text;
+using System.Threading.Tasks;
+using System.Transactions;
 
 namespace CrudOperations.Business.Concrete
 {
@@ -16,16 +19,42 @@ namespace CrudOperations.Business.Concrete
             _categoryRepository = categoryRepository;
         }
         
-        public Category CreateCategory(Category category)
+        public async Task<Category> CreateCategory(Category category)
         {
-            //try catch log
-            return _categoryRepository.CreateCategory(category);
+            try
+            {
+                var nameExist = await _categoryRepository.IsNameExistAsync(category.Name);
+                if (String.IsNullOrEmpty(category.Name)) // fluent validation kullanılması daha sağlıklı olur
+                {
+                    throw new Exception("boş isimli kategori oluşturulamaz");
+                }
+                if (nameExist)
+                {
+                    throw new Exception("bu isimde kategori zaten var");
+                }
+                return _categoryRepository.CreateCategory(category);
+            }catch(Exception ex){
+                throw new Exception(ex.Message); // bir yerlere loglanmalı
+            }
         }
 
-        public void DeleteCategory(int id)
+        public async Task<bool> DeleteCategory(int id)
         {
-            //try catch log
-            _categoryRepository.DeleteCategory(id);
+            try
+            {
+                var relCat = await _categoryRepository.HasRelatedCategoryAsync(id);
+                //aynı kontrol kategoriye bağlı ürünler için de yapılmalı 
+                if (relCat)
+                {
+                    throw new Exception("bu kategoriye bağlı alt kategoriler olduğu için silme işlemi yapılamaz");
+                }
+                _categoryRepository.DeleteCategory(id);
+                return true;
+            }
+            catch(Exception ex) 
+            {
+                throw new Exception(ex.Message);
+            }
         }
 
         public List<Category> GetAllCategories()
@@ -40,10 +69,25 @@ namespace CrudOperations.Business.Concrete
             return _categoryRepository.GetCategoryById(id);
         }
 
-        public Category UpdateCategory(Category category)
+        public async Task<Category> UpdateCategory(Category category)
         {
-            //try catch log
-            return _categoryRepository.UpdateCategory(category);
+            try
+            {
+                var nameExist = await _categoryRepository.IsNameExistAsync(category.Name);
+                if (String.IsNullOrEmpty(category.Name)) // fluent validation kullanılması daha sağlıklı olur
+                {
+                    throw new Exception("boş isimli kategori oluşturulamaz");
+                }
+                if (nameExist)
+                {
+                    throw new Exception("bu isimde kategori zaten var");
+                }
+                return _categoryRepository.UpdateCategory(category);
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.Message); // bir yerlere loglanmalı
+            }
         }
     }
 }
